(fix) Poll the video until the process is complete
In some cases, the mp4 can takes a really long time to be encoded.
Since we stoped to poll the video as soon as it is ready to stream, this
could lead to unplayable videos in browser that do not support dash or
hls.

This commit changes this by added a new flag, completed, that will mark
the video as being "fully" ready. We also continue to poll the video
until the process is 100% done.